+++ RELEASE-NOTES.md
@@ -1,5 +1,10 @@
 # Release Notes:
 
+## v2018.04.1
+
+- Add FingerprintUtilitiesService.segmentSlap() and FingerprintUtilitiesService.convertWsqToImages().
+- Add few failure codes.
+
 ## v2018.03.2
 
 - Add InitializeResponse.FailureCodes.DRIVER_NOT_INSTALLED.